Remove bin/scripts duplication
If a script does not need to be modified for a pathogen ingest, pull
script during runtime instead of maintaining a potentially diverging
copy.

Use a permalink for each script to allow us to version the software we
use in this workflow without being affected by upstream changes until
we want to bump the version. This design adds more maintenance to this
workflow, but it also protects users against unexpected issues that are
outside of their control.
Discussed in https://github.com/nextstrain/ebola/pull/6#discussion_r1048835183

However, this commit gets reverted later based on discussions.