Ingest: Using Snakemake modules for ingesting data into the Nextstrain build system
The running behavior is:

* If `data` directory exists, proceed with build as normal.
* If `s3_src` is pointing to a pre-cleaned dengue dataset, fetch dataset and proceed as normal.
* Otherwise, fetch dataset using ingest, clean it, and proceed as normal.

In order to implement the above, we are using Snakemake modules¹
which allows us to define a `prefix` for the ingest input/output paths.
However, these do not get applied to the scripts called within `shelL` .
So we are using this workaround² to add a base dir to the scripts.

¹ https://snakemake.readthedocs.io/en/stable/snakefiles/modularization.html#modules
² https://stackoverflow.com/a/66890412

Co-authored-by: Jover Lee <joverlee521@gmail.com>