ingest/nextclade: Set final targets
Set final targets to be the pair of metadata_denvX.tsv and sequences_denvX.fasta files.

The insertion, alignment and translation files had been produced by nextclade
in other pathogen repos, however, we are not using them in dengue yet. If we need them
later on, we can add in alignment and translation files (split by serotype) back in
at that time.