Use wildcards conditionals to further combine rules
Use wildcard conditionals to further combine rules that take a different
set of input files for the genome and E gene builds.

Case 1: The E gene build excludes the "augur clades" command, resulting in the
absence of a clades_{serotype}_E.json file in the E gene export rule.

Case 2: The reference Genbank files for E genes are dynamically generated
in the results folder.