Exclude withdrawn and duplicated records
Duplicates, referring to identical sequences that may or may not be distinct samples, were highlighted in the following comment: #28 (comment). Additional discussion can be found in the thread starting here: #28 (comment).

It is crucial to note that some of these excluded duplicates actually represent patents (PAT) for vaccine candidates, and as such, they are omitted from the phylogenetic analysis of current dengue diversity.

When encountering duplicates and there is a reference sequence identified with prefix (NC_), the preference was to retain the reference and exclude other duplicates.

In cases where multiple VRL records share the same nucleotide sequence, the earliest sample in alphabetical order was selected, and the others excluded.

The rationale for each exclusion is documented in the respective comments in the exclude.txt file.