Add color orderings
The "Nextclade genotype" annotation was being displayed out of order in the
color scheme. This commit adds a color ordering file that can be used to specify
the order of colors for nextclade_genotype, country, region, and recency.

This was copied and modified from the "colors" rule in RSV's workflow

* https://github.com/nextstrain/rsv/blob/a1788ce2c9c4375fb5a06d1426c64c45cf90225f/workflow/snakemake_rules/export.smk#L13-L27