Fix serotype annotation for MT597439
MT597439 incorrectly annotated as /organism="dengue virus type 2" in NCBI
but should be /organism="dengue virus type 4" based on the submitter publication
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7952331/ where they correctly assign
it as DENV4/II (See Fig 1b, sample 43257 highlighted in yellow).

Separately, someone will be requesting that NCBI to correct this entry. But this commit
manually corrects the serotype annotation in the annotations.tsv file in the meantime.

Update exclude file since MT597439 no longer divergent in the denv4 tree