Infer ancestral root in the phylogenetic workflow (#61)
This commit reverts the change introduced in https://github.com/nextstrain/dengue/pull/57/commits/94316551c8af6270134411305098a0694eb17811
which set the REFSEQ references as the roots in the auspice trees.
Turns out this is not the desired behavior for the dengue phylogenetic trees.

Instead, we want to use the sequences and metadata to infer the ancestral
root of each tree. These inferred-ancestral roots will then be used in building
the nextclade datasets as hard-coded reference and root of the scaffold trees.