phylogenetic: Add multiple input support
Updated the filepaths for the rules copied from pathogen-repo-guide 
to support the serotype wildcard.

Based on discussion in https://github.com/nextstrain/dengue/issues/114, 
I'm keeping the status quo of using the serotype wildcard in inputs but in the 
future we will migrate to a single input with filtering by serotype. That 
requires a major overhaul of the config schema that I don't want to hold up 
the multiple inputs feature.